+++ classes/actions/ActionTopic.class.php
@@ -322,7 +322,7 @@ class ActionTopic extends Action {
 		/**
 		 * Заполняем поля для валидации
 		 */
-		$oTopic->setBlogId(getRequestStr('blog_id'));
+		$oTopic->setBlogId((int)getRequestStr('blog_id'));
 		$oTopic->setTitle(strip_tags(getRequestStr('topic_title')));
 		$oTopic->setTextSource(getRequestStr('topic_text'));
 		$oTopic->setTags(getRequestStr('topic_tags'));
@@ -466,7 +466,7 @@ class ActionTopic extends Action {
 		/**
 		 * Заполняем поля для валидации
 		 */
-		$oTopic->setBlogId(getRequestStr('blog_id'));
+		$oTopic->setBlogId((int)getRequestStr('blog_id'));
 		$oTopic->setTitle(strip_tags(getRequestStr('topic_title')));
 		$oTopic->setTextSource(getRequestStr('topic_text'));
 		$oTopic->setTags(getRequestStr('topic_tags'));
